Refactored the 'Apriltag' class by migrating methods which interface directly with the Limelight into 'Limelight.java.'
