+++ src/main/java/frc/lib/util/AprilTag.java
@@ -10,11 +10,10 @@ import edu.wpi.first.math.geometry.Rotation3d;
 import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.networktables.NetworkTable;
-import edu.wpi.first.networktables.NetworkTableEntry;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import frc.robot.Constants;
-import frc.robot.RobotContainer;
 import frc.robot.commands.AdvanceToTarget;
+import frc.robot.subsystems.Limelight;
 
 // The AprilTag class, harnessed in order to
 // calculate and execute various 'vision'-related
@@ -23,59 +22,39 @@ import frc.robot.commands.AdvanceToTarget;
 
 public class AprilTag {
 
-    public static final NetworkTable TABLE = NetworkTableInstance.getDefault().getTable("limelight");
+  public static final NetworkTable TABLE = NetworkTableInstance.getDefault().getTable("limelight");
 
-    public final AdvanceToTarget advanceToTargetInstance;
+  public final Limelight limelightInstance;
+  public final AdvanceToTarget advanceToTargetInstance;
 
-    public AprilTag(AdvanceToTarget advanceToTargetInstance) {
-      this.advanceToTargetInstance = advanceToTargetInstance;
-    }
-
-
-    // Fetches and returns the NetworkTable entry of the parameter provided in the method call.
-
-    public Optional<NetworkTableEntry> entry(String key) {
-        return Optional.of(TABLE.getEntry(key));
-    }
-
-    // Executes periodically while quering the primary target apriltag's 
-    // current position and printing the distance to such in the console.
-
-    public void aprilTagPeriodic() {
-      Optional<Transform3d> targetPosition = targetPos();
-
-      if (targetPosition.isPresent()) {
-        // System.out.printf("Target position: {x: %.3f, y: %.3f, z: %.3f}\n", t.getX(), t.getY(), t.getZ());
-        if (getDirectDistance(targetPosition).isPresent()) {
-        //   System.out.println("Current spatial distance to primary target: " + getDirectDistance(targetPosition).get());
-            determineTargetRotationalOffset(Optional.of(targetPos().get().getTranslation()));
-            advanceToTargetInstance.specifyTarget(targetPosition.get());
-        }
-      } else {
-        System.out.println("No Limelight target.");
-      }
-    }
-
-  // Fetches, calculates, and returns the offset from the robot to the current primary apriltag target.
+  public AprilTag(Limelight limelightInstance, AdvanceToTarget advanceToTargetInstance) {
+    this.limelightInstance = limelightInstance;
+    this.advanceToTargetInstance = advanceToTargetInstance;
+  }
 
   public Optional<Transform3d> targetPos() {
-        // Optional<double[]> targetPose = Optional.of(entry("targetpose_robotspace").get().getDoubleArray(new double[0]));
-        Optional<double[]> targetPose = Optional.of(new double[] {-5, 0, 0, 0, 0, 0});
-        
+    return limelightInstance.targetPos();
+  }
 
-        if (targetPose.isPresent()) {
-          return Optional.of(new Transform3d(
-              new Translation3d(targetPose.get()[0], targetPose.get()[1], targetPose.get()[2]),
-              new Rotation3d(targetPose.get()[3], targetPose.get()[4], targetPose.get()[5])
-          ));
-        } else {
-          return Optional.empty();
-        }
-    };
+  // Executes periodically while quering the primary target apriltag's 
+  // current position and printing the distance to such in the console.
+  public void aprilTagPeriodic() {
+    Optional<Transform3d> targetPosition = limelightInstance.targetPos();
+
+    if (targetPosition.isPresent()) {
+      // System.out.printf("Target position: {x: %.3f, y: %.3f, z: %.3f}\n", t.getX(), t.getY(), t.getZ());
+      if (getDirectDistance(targetPosition).isPresent()) {
+      //   System.out.println("Current spatial distance to primary target: " + getDirectDistance(targetPosition).get());
+          determineTargetRotationalOffset(Optional.of(limelightInstance.targetPos().get().getTranslation()));
+          advanceToTargetInstance.specifyTarget(targetPosition.get());
+      }
+    } else {
+      System.out.println("No Limelight target.");
+    }
+  }
 
   // Calculates the distance between the robot's current estimated 
   // position and the one of the primary target apriltag.
-
   public Optional<Double> getDirectDistance(Optional<Transform3d> targetVector) {
 
     if (targetVector.isPresent()) {
@@ -85,7 +64,7 @@ public class AprilTag {
     }
   }
 
-  public Optional<Double> getDualDimensionalDistance(Optional<double[]> originPosition, Optional<double[]> targetPosition) {
+  public Optional<Double> getPlanarDistance(Optional<double[]> originPosition, Optional<double[]> targetPosition) {
     if (originPosition.isPresent() && targetPosition.isPresent()) {
       return Optional.of(Math.sqrt(Math.pow(targetPosition.get()[1] - originPosition.get()[1], 2) + Math.pow(targetPosition.get()[0] - originPosition.get()[0], 2)));
     } else {
@@ -105,17 +84,16 @@ public class AprilTag {
 
   // Determines the robot's position on the playing 
   // field, through the use of limelight's vector distance API.
-
   public Optional<Transform3d> determinePosition() {
-    Optional<Transform3d> relativePositionalData = targetPos();
+    Optional<Transform3d> relativePositionalData = limelightInstance.targetPos();
 
-    Optional<Long> aprilTagIdentifier = Optional.of(entry("tid").get().getInteger(-1));
+    Optional<Long> aprilTagIdentifier = Optional.of(limelightInstance.entry("tid").get().getInteger(-1));
 
     if (aprilTagIdentifier.isEmpty() || aprilTagIdentifier.get() > 16) {
         return Optional.empty();
     }
     
-    Transform3d currentAprilTagLocation = Constants.APRILTAGS.get(aprilTagIdentifier);
+    Transform3d currentAprilTagLocation = Constants.Limelight.APRILTAGS.get(aprilTagIdentifier);
 
     double[] currentRelativeLocation = {
         currentAprilTagLocation.getX() + relativePositionalData.get().getX(), 
@@ -147,13 +125,8 @@ public class AprilTag {
   // Calculates the relative rotational offset to the 
   // given "targetPosition," based off of the current 
   // orientation of the robot within 3D space.
-
-public Optional<double[]> determineTargetRotationalOffset(Optional<Translation3d> targetPosition) {
-        // Optional<Transform3d> currentRelativePosition = determinePosition();
-        Optional<Transform3d> currentRelativePosition = Optional.of(new Transform3d(
-            new Translation3d(0, 0, 0),
-            new Rotation3d(0, 0, 0)
-        ));
+  public Optional<double[]> determineTargetRotationalOffset(Optional<Translation3d> targetPosition) {
+        Optional<Transform3d> currentRelativePosition = determinePosition();
 
         if ((targetPosition.isEmpty() || currentRelativePosition.isEmpty()) && !validTargetInput(targetPosition)) {
           return Optional.empty();
@@ -163,7 +136,7 @@ public Optional<double[]> determineTargetRotationalOffset(Optional<Translation3d
         double yAxisOffset = currentRelativePosition.get().getY() - targetPosition.get().getY();
         double xAngularOffset = (Math.atan2(yAxisOffset, xAxisOffset) * (180 / Math.PI)) - currentRelativePosition.get().getRotation().getX();
       
-        double yAngularOffset = Math.asin((targetPosition.get().getZ() - currentRelativePosition.get().getZ()) / getDualDimensionalDistance(Optional.of(new double[] {currentRelativePosition.get().getX(), currentRelativePosition.get().getY()}), Optional.of(new double[] {targetPosition.get().getX(), targetPosition.get().getY()})).get());
+        double yAngularOffset = Math.asin((targetPosition.get().getZ() - currentRelativePosition.get().getZ()) / getPlanarDistance(Optional.of(new double[] {currentRelativePosition.get().getX(), currentRelativePosition.get().getY()}), Optional.of(new double[] {targetPosition.get().getX(), targetPosition.get().getY()})).get());
 
         if (Double.isNaN(xAngularOffset) || Double.isNaN(yAngularOffset)) {
           return Optional.empty();
