+++ src/main/java/frc/robot/Constants.java
@@ -62,11 +62,6 @@ public final class Constants {
     public static final double launcherHeight = 0.574;
     public static final double feedVelocity = 0.5; // just a placeholder value - will need to be changed
   }
-  public static final class Vision {
-    public static final double kLimelightAngleDegrees = 55;
-                                                      //height  //conversion const
-    public static final double kLimelightHeightMeters = 16 * 0.0254;
-  }
 
   public static final class Vision {
     //all units in meters
@@ -75,12 +70,6 @@ public final class Constants {
     public static final double LimelightOffsetZ = 10.5 * 0.0254;
     public static final double LimelightOffsetX = 0.32;
     public static final double LimelightOffsetY = 0.0;
-
-  }
-  public static class Launcher {
-    public static final double gravityAcceleration = 9.81;
-    public static final double launcherHeight = 0.574;
-    public static final double launcherWheelRadius = 0.04826;
 
   }
 
@@ -230,16 +219,7 @@ public final class Constants {
   public static final int kDriverControllerPort = 0;
   public static final int pcmCanID = 15;
 
-  public static class Launcher{
-    //Unsure if we will need this but keep in case the Launch Angle calculations launch the note towards the AprilTag and not the hole
-    public static final double ampHeight = 0.66;
-    public static final double speakerHeight = 1.984;
-    public static final double trapHeight = 1.436;
-    public static final double launcherHeight = 0.574;
-  }
   public static class Limelight {
-
-
 
   /**
    * A key-value map, mapping each AprilTag to its position on the 2024
